+++ Source/Mosa.DeviceSystem/RestartService.cs
@@ -32,7 +32,7 @@ namespace Mosa.DeviceSystem
 			// TODO: Get Reset information (type, address, value) from ACPI
 			int type = 0;
 			byte address = 0;
-			uint value = 0;
+			byte value = 0;
 
 			// If via PCI Bus, get Host bridge controller:
 			{
